Added the initial core-state functionality which is used by the scene-tree expansion to insert itself into the core-state.
Fixed the scene-tree expansion to understand how to record the
reinitialize-instance into the core-state.

To run the code:

,re
(asdf:load-system :gear)
(in-package :gear/example)

cut and paste the (defun expansion () ... ) and the testme function after it
into your repl from example/scene-tree.lisp.

(testme) -> a three item list. The first is the scaffolding hash table
for the scene tree (debugging), the secon is the actor root object in the
scene tree, and the third is the core-state instance.