Fixed the camera component to behave either as a regular old camera that looks down the neg :z with an up of :y and a right of :x of the actor's transform, or as a "look at this actor" camera. In the latter, the camera moves as the actor does, but the camera component computes its own view as required to trsack a different object in the center of view.  The camera is NOT a follow camera, like a bird on a shoulder. It is insteasd a tracking camera, like a security camera at a mall rotating to view someone as they pass.
The camera can easily switch between tracking something or just being a
regular camera.

Changed the demo scene to demonstrate the tracking nature of the camera.
Fixed camera near/far clip to be much larger. I think we should have a near
far for the ortho camera and a near/far for the perp camera, since they
are not equivalent. Added a function that will set the camera-component to
track a particular actor. Exported it from gear so the user can use it.

Changed the clear-color to be less noticable. It actually DID help me
debug a shader problem!

Iterated the unlit-texture shader to interpolate vertex colors. Will
soon add actual texturing.