Beginning work on Step 4. The component type ordering.
So far, I've included cl-graph and split-sequence and wrote a few utility
functions. I haven't yet gotten to the graph-definition form parsing yet.
The DSL design leans itself more towards traditional parsing techniques
as oopsed to the scene dsl, which required large rewrites and a macro
to deal with user executable forms. So far in the graph dsl, there are no
user executable forms, so I can get away with not needing to actually embed
this into a macro.