Serious work on Step 4.
Starting at the roots of the graph-definition, the splices are
expanded in the graph when the splices are subdags in the same
graph-definition. Currently, only splices that are not specified in a
:depends-on form are expanded. I need to add that latter feature in a later
commit. This commit produces a bit of output with a toposort at the end
of the type dependency graph. However, this code is incomplete.

Also, notice that splice expansions are scoped incorrectly. That needs
to be fixed also. Well, one step at a time....