Checkpoint commit on the way to adding texture descriptors and allowing materials to use them.
Broke the texture class up into an image class and a texture class. The
latter is not used in this commit, but will be in the next commit
after some refactoring in texture.lisp and materials.lisp.

The image class is currently hard coded to use TGA images. It should
be straight forward to extend it if we need to later.