Demonstrate to mfiano that his sprite-sheet method works in FL.
Implemented a 'sprite-sheet' component in mfianos-tests project
(along with shaders and materials).  This new component uses a shader
to construct geometry into which a sprite is picked out of a sprite
sheet to be rendered onto it.  The sprite-sheet code is unpolished.
The sprite-sheet also has rudimentary animations in which you can
specify a duration for a periodically repeating sequence of sprites
in the sprite sheet. Each sprite-sheet component can have more
than one animation and you can choose between them independently
for each component instance. Currently, each component just has
a single animation of a number of cells.  The demo.scene for this
shows a constructed ship with exhaust in the canonical manner of
representation. THe demo.scene is very instructive.

Added FRAME-TIME function for export out of fl.core that uses the
context instance.

Flipped the origin of the TGA to be bottom left so it rendered right.
The TGA origin and relation to the spec file is slightly tenuous.

I discovered the shared storage is somewhat fucked, so we need to
rethink how that works again. So, we remake vaos, vbos, ssbos and
other stuff unecessarily for each sprite-sheet instance. In fact,
the ssbo binding point of 8 is truely screwed since each instance
reuses it since it was hard coded. Luckily, the data _happens_ to
be the same for all of them, so it works. But, the concept of the
binding point for ssbos desperately needs cleanup.