Reordered how the calls to shadow work in the sprite-sheet. This better indicates how the work flow of using them works.
Fiirst, we create-a buffer suitable for a particular interface block,
and then, right before use, we perform the binding of that interface
block and the buffer. Of course, we have the option for a permanent-ish
binding point if we moved the interface block binding to the
initialize-component method. And, sometimes that'll happen, but sometimes
it won't.