Pushed the materials functionality to code completion, minus inteface blocks. I put all of this into the psilord project, except where I needed to fix code in FL to make it work.
Now, you can COPY-MATERIAL a material (deep copy it) and assign it a new name.
You can also use MAT-UNIFORM-REF to get the semantic value of
a uniform, and MAT-COMPUTED-UNIFORM-REF to get the computed value of the
uniform (the value very nearly ready to be sent to the shader).
You can also (SETF MAT-UNIFORM-REF) a new semantic value, which will
be auto converted to the computed value.
You can supply, in the define-material itself, on the tuple for the
uniform name and value, a third entry for a function that will be placed
first in the computation sequence from semantic value to computed value,
allowint you to put whatever the hell you want into the semantic-value
entry for the uniform, as long as your function convertes it down to a
meaningful type that can pass through the rest of the compositions to
end up a computed-value.