Use a thread safe queue for recompilation tasks (like shaders) in FL. When we stopped running FL in sdl's main thread, it became harder for C-c recompilation to work. This is because slynk will fire off a worker-thread to perform the evaluation of the form from emacs. That means, in the case of shader recompilation, that shadow eventually called the shader-modify-hook from a different thread than FL is executing (and the FL thread owns the opengl context).
So now, in core-state there is a thread safe recompilation queue. And, when
we use the "hook" idiom to recompile something (like shaders), that function
must enqueue a work packet into the recompilation-queue in core-state.
Then, later, in core.flow, there is a special and new flow-state called
'perform-recompilations' which happens first in the work to compute the frame.
It is there we actually dequeue the work packets off and shuttle them to
the appropriate functions to perform the work described in them.

This system fixes three problems:
A) Different threads requesting recompilation of thing that may involve opengl.
B) It codifies a pipeline that any recompilation tasks should go through
   to make it maintainable and centralized in the source.
C) We can schedule exactly when we want those recompilation to happens, and
   even change their ordering if we so desire, etc, etc.