Shutdown the shader programs when the engine exits.
This fixes a bug where when we initially insert the shader-modify-hook
into shadow, it has a closed over core-state in it. This is wo we can
emit messages about shader recompiles. However, then the engine exist,
it is STILL closed over. So, when we start the engine again, shadow
uses the OLD core-state's shader-modify-hook to emit piles of
recompilation messages before the newly brithing engine overwrites it with
the new core-state reference. So, the fix is to (constantly nil) the
shader-modify-hook wrt shadow when we shut down the engine.