Update mfiano's project to use my new sprite sheet component.
This component is much sexier in that it understands how to deal with
different spec files and textures and what not, but yet keep everything
cached and shared as appropriate. EXCEPT a small piece of internal data
it generates that I haven't gotten around to sharing.

Also tweaked some stuff in movement to add a mouse fire option.