Fix the codes to emit shots like how one expects. There is a huge TODO in here which must be resolved in gamebox-math. Basically quat:from-mat4 needs a keyword argument to notify it if it should make a copy of the basis vectors, then normalize them, and then use those copies in the algorithm. Basically, quat:from-mat4's conversion algorithm EXPECTS normalized basis vectors, but it is most often the case in FL that they won't be. So, by default it should temporarily normalize the basis vectors it uses for its own conversion routine.
I also changed the angle the atan produces out into a rand of 0 <=
angle < 2pi for easier debugging of the rotation matricies of the
ship. There was nothing wrong with it mathematically, except it
was hard to see the bouncing of the basis vector values across the
boundary when debugging.

Changed the fire button to :right-shoulder since it is too hard to use :a
when actually controlling the joysticks.