+ Implemented the "annotation and override" feature.
+ Implemented a single annotation in fl.annotations called 'material'.

This annotation also designates getter/setter functions that (for
a slot annotated with this annotation) will automatically convert
the material symbol to a real material when we create a component
or setf a material symbol to its slot.  We don't have to do that in
INITIALIZE-COMPONENT anymore or do the lookup by hand at runtime.
We can also setf a true material instance to the annotated slot and
it'll do the right thing. When we read the annotated material slot, we
should get back a real material instance.  Performance is maintained
because I used the MOP so we ONLY pay a cost on actually annotated
slots in a component.

+ The material getters/setters inplement an "override form" that looks like
this in a pseudo lambda-list kinda of way (which must be quoted):

(<defined material sym> <new material name sym> &rest initargs)

and then the initargs can contain
:shader <shader sym>
:uniforms ( (:auniform-name-1 42) (:buniform-name-2 44) ... )

This allows you to change the shader (but the uniforms MUST match the new
shader you've changed to) and also to make in-place changes to any uniform.
Blocks are not supported yet.

This override form can be used in the scene DSL AND/OR in runtime code.
So, in *runtime* code in a component, you could do:

(setf (material component) `(fl.foo:base-material new-mat-name-sym :shader foo))

and it'll automagically perform the copy of the base material, name the new
copy 'new-mat-name-sym' and alter the shader to a different one.

Because of the way I implemented the transformation pipeline of the getters
and setters, theyh often (but definitely not always) can be the same
transformation function. This is reflected in the fl.annotations:material
annotation example.

+ Annotations can be forward declared (simply by use in the :annotation
list on a slot). They get a default identity functions for their
setters/getters a forward declaration situation. So, if there are never
actually any getters/setters defined, they effectively become trackable
markers on the component's slot. However, if they are declared with
DEFINE-ANNOTATION (before or after their use in a define-component)
then the actual setter/getters will be used properly at runtime.

+ All annotations end up in the new metaclass slots in the COMPONENT
base class type. So we have a record of all possible annotations after
we load the project. The user can define their own annotations with their
own behavior. The entire list of annotation after load of the project is
available without running the project.

+ There can be multiple annotations per slot in a component.

+ The invocations of annotations at runtime are optimized at compile
time (and C-c time) to provide good performance. It could be optimized
more, but what I have now is good enough.

+ We keep track of all defined component names in the project
(from any source, FL or otherwise) in the metaclass slots on the
base COMPONENT type. For those components with annotated slots, we
store all (slots, initargs, and type of annotations) on those slots,
for each component in the COMPONENT classes meta-class slots. This
will be incredibely useful when we work on and complete the "track
your resources" feature to enable we only load textures we need,
or store data/image/audio/etc files actually used by the game, as
opposed as just existing in the project directory. The entire database
of components with annotated slots is avsailable without running
the project.

+ Annotations for FL are kept in src/annotations.lisp and in a new
package which exports the symbols used with DEFINE-ANNOTATION.

+ We support C-c on DEFINE-COMPONENT and it'll update all annotated
slots correctly.

+ We support C-c on DEFINE-ANNOTATION and it'll update the annotation
db correctly.

+ Converted all projects to use the fl.annotations:material annotation
in their component slots as appropriate and tested each one.

+ Specifically, projects/mfiano/noise-test now uses a SINGLE noise material
and identity noise shader, that in the define-scene is varied using the
override form for the material.

+ Multiple annotations per slot are always executed left to right.
When setting, they take the value that is being set, the shove it though
the pipeline and then actually set the result into the slot. In the getters
they take the value from the slot, and run it through the same pipeline
in the same way to produce the result that is given back to the user.