Raw example code for generating an unrealized procedural texture.
If you run fl.example:texture-test, the marble texture is procedural.
Added a new quad in that scene to demonstrate this.

This texture is noticed as being unrealized in the PROLOGUE, then
parameters specified and data created and uploaded in terms of opengl
to the GPU.

It takes a while to start cause black-tie is not optimized so
texture generation on the >C<PU is slow.

The API in the example is very raw and will go through heavy evolution
once all the functionality for :procedural textures are completed.

I already modified some of the API since the last checkin, more to come.