Add a feature to invoke on-attach-component & on-detach-component. Each component now keeps a queue of the actors that it attach to or detached from. Just before on-update-component, all of these events in the queue are processed at once in type order for the components. Also, just before on-destroy-component is called for a component, all queued attach/detach events are invoked.
Note: The attaching/detaching concept is not a first class citizen in
the flow in the sense that we don't lockstep the attachment or
detachment process itself. This means that on-initialize-component is
called BEFORE the attach/detach events and the actor is the current
actor always for on-initialize-component. Then, when the component
is destroying, the detachment happens AFTER the destruction, hence will
never get called (and is left to the garbage collector). And, if you
wanted it called, it would be called with a missing actor slot in the
component.

We can revisit exactly what happens at the start/end of the lifetime
of a component as we use the system and discover what we really need it to
do.