Fix: Compiler error with new component events.
You shouldn't specialize the ACTOR argument for
ON-COMPONENT-ATTACH/DETACH, for a few reasons:

1. It's a compile error, because the actor class is not yet defined at
the time component.lisp is compiled.

2. The actor instance is passed by internal FL code, and will always be
an actor instance.

3. Because it will always be an actor instance, it is pointless, and
will only result in a slower method if it is specialized on an actor
class.

4. It propagates to user code, if they look at example code and do the
same.

In short, specializing on an actor class instead of handling the
argument unspecialized, forces code reorganization, runtime performance
cost, and user code verbosity.

Therefor, I'm unspecializing the actor argument from these 2 component
event methods.