Fixed all transform rotations to be local axis by default, not world axis.
The flm:rotate now takes a :local or :world for quaternion rotations.
The difference is the order in which the rotation is applied to the
input quaternion.

I did the same thing for the matricies, although it is not entirely
complete since we don't use those interfaces very much. I left out
the type specializations for the ones we don't use and we can fill them
in later if we actually do use them. This kept the change to be pretty
minimal.

The only scene which changed as a result of this was damaged helmet
since our other scenes were too simplistic to have been affected by it.

The change done to damaged helmet is now after the initial rotation by
pi/2 around the +x axis, now we do -0.6 around the locally transformed
z axis to get it to rotate in the manner expected--as opposed to the
.6 +y _world_ axis rotation it was doing before.