Make the missing-material use debug-texture.
When I'm debugging something I don't want to see magic magenta...I could
mistake that for bad vram or shader or anything really. If a texture is
missing, show something meaningful like the debug-texture as that is
usually what you want for debugging uv's etc anyway when loading a new mesh.