Removed internal fl.util and fl.math
This is because the utilties we use are solid and won't change except
for the ocassional accretion, but no changes to existing utility
functions.

As for the math, that is stable and shouldn't really change at all
unless we discover incorrect operations again, but I don't think that
will be the case.

It turns out, both of these projects are very useful outside of the
context of first-light, and maintaining two copies, merging changes back
and forth is a hassle, since they are both used by myself and others
outside of first-light.

All other first-light modules are very domain-specific and will remain
as such. These two libraries don't make a lot of sense to be internal
since they are used in the wild, and changes need to be back and forward
ported to support that which is more work for little benefit considering
their stability.