Added TRANFORM-POINT, TRANSFORM-VECTOR, TRANSFORM-DIRECTION.
These use a TRANSFORM component to affect a vector in various ways.

TRANSFORM-POINT will transform a vector with rotation, translation, and scale.
TRANSFORM-VECTOR will transform a vector with rotation and scale.
TRANSFORM-DIRECTION will transform a vector with rotation only.

I need to write the inverses, which isn't that hard, I just haven't done it.