Completion, for now, of the entire collider-system. Good for a game!
+ The collider system can devolve to n^2, but if you have a decent number of
intelligible physics-layers and your colliders are sanely distributed through
them, you'll get a lot of mileage out of the system.

+ Added a shader to visualize the collider/sphere. It will visualize as a spiral
wound around the sphere--a parameterized sphere. This was ridiculously fast to
write and I have limited time. The visualization is simply a special procgen
geometry built from a material and shader called collider/sphere. I put this
into core, cause it is meant to be there and used by many different projects.

+ Changed the dispatch path of the collision events. Now, all collision events:
enter/continue/exit are routed to the collider component itself and the collider
component is responsible for routing it to the same on-collision-* interface on
the referent, if any. This allows us to have the collider keep secret state
about itself (like if it is in contact with anything and should alter its
visualization). In the future the functionality in the collider will likely
grow.

+ Added and removed a bunch of debugging.

+ Added the "collision-simple-destroy-0" prefab which will demonstrate the
collider system working. Since I've committed that the visualize slot in the
collider/sphere defaults to T, you'll see the colliders in the game and there
will be some repl output when they collide.

+ Currently all collisions are discrete and there are no spatial data structures
helping out the collisions.

+ The referent of a collider may be NIL. The referent must be a component
that implements the on-collision-enter, on-collision-continue, and
on-collision-exit protocol.

+ The on-layer of a collider may NOT be NIL. This is a rough edge, it should
runtime default to something (which will always be wrong, but legal in the
collider-system).