Large initial commit of lisp gam jam april 2019 codes.
This uses no example components, but definitely FL components.
All components for this game beyond those included in FL are written into
this file. This game will only ever be a single file.

Using a gamepad, you can move with the left stick, rotate with the
right stick, and fire with right shoulder. No enemies yet.

I need to add a collider to the player ship, and uh, enemies and such.

Other than not having font support, which I can fake later, this is
progressing along nicely.