Fixed a showstopper bug with FL:REF wrt :copy/:link.
Each prefab has a lexical environment to support the FL:REF call in the value
position of all components that exist in that prefab.  The problem was that when
you :copy or :link another prefab, the components flowed from that prefab into
the new one--and more complex, each component's init arg could potentially stay
the same or be replaced and it moved through the :copy/:link forms in deeper and
deeper nests of prefabs!

So now, a component initarg value is thunked into an injectable-ref-value-thunk
instance which also includes the pandoric accessor to the specific closure that
value is valid under. Then, later, we can poke the correct reference information
into that closure just before we force the thunk and produce the actual value.

We COULD put an INJECT-REF-ENVIRONMENT around EACH value form, but that would
produce more garbage than necessary. Using one for the prefab as a whole just
means that we'll reset some of the values to the same thing a few times while
making the instance of the prefab, but that is prolly cheaper than making more
garbage.