Update the gamepad-db.txt file.
Fix the script to put the file into the right place since it moved
eariler.
Fix the script to add custom definitions. Currently there are none, but
only because the canonical version JUST made a new release that added by
default the one I was just about to put in here. So might as well just
keep the code I wrote.