mfiano suggested changed to the asteroids. This is some of them.
All 6 asteroids are now able to be randomly chosen by the asteroid field.

The asteroids (and their colliders) are randomly, but uniformly, scaled
in beterrn .75 and 1.25 from their original size.

I have not yet implemented adjusting the "rotation" since that is a function
of slowing down or speeding up the duration of the animation and I have not
yet inspected the ACTION component to implement that.