First pass at a tags component which is a candidate for core.
This component associates tags, which are symbols, to the actor
which contains the TAGS component. This allows fast lookup to see
if an actor has a particular tag, or what are the actors that have
a specific tag. The code seems to work via inspection of the data
structures during a running game. But it needs more testing and
some more API work to make it easier to use.