First commit.
This is a first pass that establishes some new PLNs, and uses them where
required.

A second pass of this refactor will follow, which will move most of the
first-light systems first-class citizens of the core library, as we need
packages in core loaded before modules are loaded to remove some circularities.