Starting to add the ability for planets to get hit and blow up.
Currently a planet can be hit a configurable number of times and then blow up.
I need to add dramatic continuous explosions in a circle around the planet
when the hit points are getting dangerously low. Prolly need to add a few more
things as I test when I notice them.