Added LEVEL-MANAGER componment.
This component is in each level and knows how to manage the
state necessary to detect win/loss conditions (such as how
many planets are alive, etc). The whole of the API to this
component is not yet completed.

The DIRECTOR looks for this component, and will use the API to it
as a means to control the level dofficulty and detect win/loss
conditions.