Added a :level-complete transitional phase when you beat a level.
Added a time-keeper component which draws a bar on the screen and
reduces its time as configured.

Added a :dangerous tag to all projectiles so we can destroy them easily.

The director processes the :level-complete state and it is possible
to run through all the levels, "beat" the game, and start over.

Added a "Level Complete" sign so you know when you beat a level.