Remove the fl.shader.user package because it was a bad idea.
Users should create their own package per project to write shaders in. All they
have to do is (:use :first-light.shader). The reason it was a bad idea is
because if someone is working on multiple projects, all of the shaders combined
occupy a single namespace. The chance for collision increases with the number of
projects and symbols representing shader function/program names. This new way,
users are in charge of their own packages for shaders, giving them whatever
flexibility they need.