Support FULL recompilation of define-texture. Even the texture type!
When you recompile the define-texture, if that texture already exists in the
GPU, it will be deleted and all resources freed. Then a new texid is created,
bound, and all images will be loaded again from disk (or wherever) and have all
new parameters applied.  The texture instance is reused so when materials are
using it in their closures, the only thing that changes is the texid so the
changes appears instantaneously.