When the player is firing, lock it to the physics rate.
This is so when the shots come out, they all come out in exactly the same
spot and travel in the same way. This removes variation from the shot
shows up and how it trasvels when they previously were emitted in the
update function. There, the shot could come anywhere between the last
physics frame to the current one and the variation was unacceptable to
actual shmup players.