Load all 120 glTF examples.
I had to do some minor fixing of when gltf type fields were nil
and I rewrite the file loader code a little and will soon change it to
use something in golden-utils. Have to do some profiling too, since it is
a little slower than I would have expected, but it might be because some
example files are huge and take a bit to parse.