Finally fixed :rotate/inc so it works properly!
This fixes the long standing bug where before when we used q:orient to produce a
rotation that was 2pi, the object would stand still. This is because when you
normalize that quaternion, you get the identity quaternion (in that
orientation), so the slerp we tried to do form the identity quaternion to the
oriented quaternion was meaningless.

For this fix, I changed the meaning of %incremental in the transform state for
the rotation to hold an angular-velocity vector instead.  This is a vector which
defines a spin (right hand rule) around itself whose radians per second are the
magnitude of the vector.

Now, for each frame, the %incremental's magnitude is multiplied by the frame
time to get the amount of rotation that has happened. Then this, along with the
direction of the angular-velocity vector stored in %incremental, is used to
construct a quaternion to be held in %incremental-delta. Later, the frame is
rotated by that %incremental-delta quaternion, producing the rotation required
in the time required.

With the angular-velocity vector, the rate of spin in per second always, and
greater then 2 pi rates are meaningful now.

I added four new functions to eh transform component API:

These make an angular velocity vector:
angular-velocity!
angular-velocity

These take that angular velocity vector and produce a quaternion that
represents it over time:
angular-velocity->rotation!
angular-velocity->rotation