Initial version of typechecking for gltf-accessor & descendants.
It turns out some of the typechecking requires actually checking the
real data. I left those out currently so I can think about what to do since
we want lazily loaded and/or streaming data eventually).

I left a trail of TODOs for the especially annoying things to check.
This is going to be a common theme as I continue the typechecking codes....