Two fixes: floating constants use f0 and angular-velocity pkg.
The origin mathlib, in order to take more advantages of SBCL's optimization was
altered to remove type instability. This means when calling constructors or
whatever the arguments must actually be single-floats. So, all floating point
constants in V and all examples were altered to have f0 on the end of them to
ensure they will be single-float types.

angular-velocity and angular-velocity->rotation (and friends) were lifted into
the origin math library. So I added a 'p' PLN for the origin.physics package and
change all calls to use that package.

I tested all examples and fixed all encountered problems.