Added actual collider math for sphere and cuboids only.
Added a test example "collision-test-2" which shows the cuboid and sphere
colliding.

Turns out the axes description of the OBB (page 101 RTCD) [the u in the OBB
struct] is a normalized basis. So we had to remove any scaling effects of the
transform's rotation matrix to implement the RTCD code.

Need to add the cuboid/cuboid intersection test. Currently it errors.

The math code is ugly and produces a lot of garbage, needs to be rewritten
with origin's viewpoint to make it clearer.