Refactored the configuration system.
This will make future cross porting of code from my other engine easier, and it
also simplifies parts of the API as well as makes the overall engine much more
efficient.

Before, we stored all configuration options in a hash table inside of context.
This meant that we were performing many hash lookups per second in the numerous
places we needed a configuration option value. It also meant we were performing
a lot of slow dynamic dispatch by looking in the core state for the context, and
in context for its options generic function accessor.

Now, there exists a new DSL, define-config, which defines a hash table for each
project with its configuration options, and also a set of corresponding global
variables which is what is actually used at runtime. The hash tables are for
inspection purposes, as well as the one time updating of the globals for the
particular project that is loaded.

These global variables are efficient true global variables, using the
global-vars portability library, and are unlike pseudo-global special variables.
The global definition forms expand into efficient implementation-specific
internals for representing a true global variable.

This new system should be much more efficient due to no hash table lookups many
times per frame to obtain a configuration option, as well as no generic dispatch
to drill into the core state to even get the hash table. There is no longer a
runtime configuration hash table. Additionally, it results in significantly
cleaner code, due to the fact that functions that require knowledge of a
configuration option no longer need to be passed a core or context argument, and
the (v::option v::context :foo) syntx is now reduced to v:=foo=. Global
variables have a noticeable syntax surrounded with equal signs to distinguish
them from other variables, similar to how constants use plus signs and special
variables use asterisks.

In short, this change results in more efficient runtime, shorter function
signatures when access to configuration is required, and a more concise
configuration option query call.