Make the transform API generic.
Each transform API function is now generic and can operate on 3 different object
types instead of just a transform:

1) A transform component to directly manipulate, as it was before this change.

2) An actor object, since it is assumed that a transform is a permanent
residents of their actor once the flows add it.

3) An arbitrary component, in which an attempt is made to access it's actor's
transform component. In the case of a "floating" component with no actor, a
runtime error is is caught to prevent unintuitive late errors.