Added SCALE-AROUND to the transform protocol and an example.
Basically, this allows you to scale any transform with respect to
any point in world space (which can be gotten from any local actor
space in the method I demonstrate in the example).

The obvious example is doing a raycast to hit a plane, then scaling the
plane's transform around that pivot. This gives the "google maps" like
interface with scaling and the mouse wheel. This code also works in 3D
just as well.