Added two special components to PtP to implement draw ordering.
These two components 'sketch' and 'delayed-render' are special to the PtP game
and are not indicative of how V will do this in the future. I was curious if I
could build a thing along side V's normal rendering system. Also I'm hacking on
PtP of late and wanted to add translucency effects. While a real solution for
the engine is forthcoming, it isn't happening quite yet and this was a night's
worth of hacking.

Now there is a render-order list in addition to the z layer list. For the
most part the game didn't change that much except specification of the
rendering-layer in a few places and changing of components in the prefabs.

Since I designed the game pre-rendering order, there is a little oddity in
that the hit-points component knows about the rendering layer. I'll fix that
in a future commit.