Make the projectiles understand if they should repeat.
Currently this is a property of the gun, which is a bit wrong.

What I need to do is break up the generic projectile system into prefabs for
each kind of projectile as appropriate. This will make the code simpler by
offloading knowledge about the projectiles (sprite anmiation, how long to play,
collider shape and size, etc, etc) into the prefabs themselves.