Turn off a cl-opengl feature and explain why in README.
The :cl-opengl-no-masked-traps cl-opengl feature was required for
some buggy intel shader compilers, but when not in those conditions
it causes a nasty performance problem. The problem is that the
floating point state is enabled/disabled for each cl-opengl call and
that adds up quite quickly. So we turn it off.

Later, cl-opengl will likely get a fix that swaps the sense of the
default cl-opengl uses, and we can remove the PUSH in the asd.