Compute normal-matrix in the render component using translation 0.
We needed to remove the translation since it was simply wrong when
using it to convert normals.

Updated my hacked-components in PtP to do this too.

Changed the normal-matrix uniform in the shader to be a :mat3 and
updated the damaged-helmet shader code to use it. We don't need a mat4
since the normal-matrix is ALWAYS a rotation matrix only, not a transform.

When we implement a m3:invert, the code in components/render.lisp
can be simplfieid more.