Fix a nasty problem where the :normal-matrix was incorrectly bound.
In WITH-MATERIAL, added a feature that if the uniform doesn't exist,
silently ignore the setting of it. This "policy" decision may have
to be implemented in some other manner.

Then, moved the setting of the :normal-matrix uniform from the body into
the uniform-bindings form of the WITH-MATERIAL macro.

Documented the WITH-MATERIAL macro to explain that setting uniforms in the
body (even though it'll end up before the draw call) will not have its
intended effect because the binding of those uniforms only happens for
those uniforms in the binding-form.