Add this to develop, but keep in flux until I can work on it.
I picked this off of a dead branch that fell farther and farther behind.
This code is effectively standalone, but I'm time starved to complete its
implementation. However, what is there now is in decent shape. When I am
less time starved and can complete it, I'll make a new branch to do that
work. I just don't want to lose this file.