Added matcaps to V.
There are two issues with this code:
1) gltf2 has uv origin in the upper left and V doesn't yet understand that.
So there is a bit of hitting with a hammer to get to display correctly.
This will be solved in a future commit.

2) When a material is copied with the annotation syntax, but then the prefab
is used more than once in another prefab, the copying doesn't happen
properly and an error is generated. The semantics for the meaning of a copied
material must be improved upon and implemented.

All matcap images were lifted from Blender Foundation and at the time they
provided a CC-0 copyright. We converted them from .exr format to .png format
using the supplied exr2png.sh script.

The matcap-lookup shader is a proof of concept about using id colors on a
surface to pick a particular material's information from a texture array.
However, all materials must have the same valence (meaning the same
set of information describing them, textures, etc, etc). No mixing
PBR materials with unlit materials unless everything has been reduced to
the same valence.