The commit makes 2 changes to our custom debugger restart:
1: There is no longer a "CONTINUE: Virality Engine" _and_ "CONTINUE:
REPL". Now, there is just the former, as the REPL
unblocker (update-repl) is called in the main game loop. It makes no
sense to have both, unlike in the past.

2: The name of the restart is no longer called "CONTINUE: Virality
Engine". Now it is called "ABORT: Skip processing the currently
executing frame.".

The reason for the same change is both for semantics, and keybinding
convenience:

2a: It was never actually "continuing" - it always skipped the
currently executing frame, and hoped that the error would be fixed by
the time the game loop got to that point on the next frame.

2b: By naming the restart "ABORT: ...", you are now able to hit the 'a'
key in the dbugger to choose that, rather than having to know which
numbered restart it is. 'a' always chooses the inner-most restart which
is named "ABORT". 'c' also chooses the inner-most restart named
"CONTINUE", but Common Lisp inserts its own restart named continue that
retries evaluating, which is not what we want.

2 (cont'd): So, You can think of it as "aborting the current frame",
rather than previously "continuing the next frames of the game loop".

In summary, there are no longer 2 custom restarts, which means time
correction isn't applied twice when invoked from the REPL. Also, type
'a' whenever the debugger is presented to just "do the
right thing".