Changed the default sense of V:TRANSLATE from inertial to local space.
Fixed all ramifications in the examples of this, except for:

example-sprite: The ship placement really IS in inertial space.

ptp: Some of it is conceptually in inertial space, some of it not.
I left tham all in inertial space and will fix later when the game
gets more of my free time.