Added a selector example to select any of the official examples.
The selector prefab is '(("example-selector" examples)) and it brings up
a menu to select other examples. There are instructiosn in the menu for
the keybindings of it. You can enter and leave any example.

While doing this change, it generated other changes all related to it.

+ Added MAP-ACTORS which maps a function across all actors in the
parent (and when given :universe and a context will do it for the
entire scene tree EXCEPT for the universe actor).

+ The material annotation form now automatically gives a new name for the
copied material based on what it was told in the annotation form.