Separated datatypes out into their own files.
This allows me to move them to a special directory so they are always loaded
first. When they move to defstruct, this will ensure all later code gets
recompiled properly when loading the system after changes to them. Structure
accessors are inlined (unless otherwise told not to, which we don't do by
default).