Renamed src/structures to src/datatype.
This folder should only contain raw defclass or defstruct definitions.
Initargs should be stuff from the dependencies for V, or set up in the
constructor function created for the datatype if it cannot be placed in
the :initform.

(And the occasional MOP thing required by the spec to exist in certain places.)