Entail Michael Fiano's Origin Math Library into Virality's source.
The purpose of this entailment is to simplify evolution, maintenance, feature
synchronization, and debugging of the math library and the connection between
the math library and Virality.

The new system has a prefix name of vorigin to distinguish it from regular
origin installs in quicklisp and such.

The specialization-store constructors for origin were previously removed
because of undefined behavior on different platforms. I had to add v3:vec*,
v3:vec*, v4:vec* constructors to convert arbitrary number types specifically to
float.  There were many simple changes like (v3:vec) -> (v3:zeros) and such to
conform to the new API.

Added a m4:*v3 to multiply a 4x4 matrix by a 3x1 vector treated like a 4x1
vector with an assumed 1 w coordinate.

Added a somehow dropped sphere collider to the ship01-interactive prefab.

Minor whitespace changes.