Implemented the MAKE-PROJECT feature.
This allows a gamedev to create a new project in a similar way to
Quicklisp's quickproject.

Notable features are that the gamedev can make their own templates and
use the subtitution patterns with them. Most everything is in the gigantic
doc string for MAKE-PROJECT.

We prolly should put the template keyword -> namestring associations into
a better place, like the equivalent of an rc file or some such.